Updated data types

+++ Cost.h
@@ -9,7 +9,7 @@ public:
 		SquaredMean,
 	};
 
-	static float GetCostOf(size_t networkOutputLength, double* neuronOutput, double* Y, CostFunction costType)
+	static double GetCostOf(size_t networkOutputLength, double* neuronOutput, double* Y, CostFunction costType)
 	{
 		switch (costType)
 		{
@@ -22,14 +22,14 @@ public:
 
 	static double SquaredMeanLoss(size_t outputLength, double* neuronOutput, double* Y)
 	{
-		float mean = 0;
+		double mean = 0;
 		for (size_t i = 0; i < outputLength; i++)
 		{
 			mean += SquaredMeanLoss(neuronOutput[i], Y[i]);
 		}
 	}
 
-	static float SquaredMeanLoss(double neuronOutput, double Y)
+	static double SquaredMeanLoss(double neuronOutput, double Y)
 	{
 		return pow(neuronOutput - Y, 2);
 	}
